==> scripts/2.4_fig_model_performance.R <==
library(tidyverse)
library(dismo)
library(cowplot)
maxent_model <- readRDS('./work/maxent_model_selection/maxent_eval_select.RDS')



# Plot part A+C

## Plot sample and background densities (A)
model_data <- bind_rows('Presence'=maxent_model@presence, 'Background'=maxent_model@absence, .id='type') %>% pivot_longer(-c(type), names_to = 'feature')
model_data$feature <- model_data$feature  %>% recode_factor(bio12='BIO12',bio15='BIO15',bio1='BIO1',bio2='BIO2', elev='Elevation', bio7='BIO7')

plt_dens <- ggplot(data=model_data, aes(x=value, fill=type))+
  geom_density(alpha=0.6)+
  facet_wrap(.~feature, scales='free', nrow=1)+
  scale_fill_manual(limits=c('Presence','Background'),
                    values=c('red','grey'))+
  labs(x='', y="Density", fill='')+
  theme_cowplot(font_size = 7, font_family = 'sans')+
  theme(text = element_text(size = 7, family='sans'),
        strip.text = element_text(size = rel(1)),
        strip.background = element_rect(fill = NA, colour = 'black', size = 0.4))

plot(plt_dens)

## Plot feature importance (C)
View(maxent_model@results)  # Just copy-paste the values because this list is difficult to transform
feature_importance <- data.frame(feature=factor(c('BIO1', 'BIO15', 'BIO12','BIO7','BIO2', 'Elevation','BIO1', 'BIO15', 'BIO12','BIO7','BIO2', 'Elevation'), levels=c('BIO12', 'BIO15', 'BIO1','BIO2', 'Elevation','BIO7'),ordered = T),
                                 indicator=c('contribution', 'contribution', 'contribution', 'contribution','contribution', 'contribution', 'permutation importance', 'permutation importance', 'permutation importance', 'permutation importance', 'permutation importance', 'permutation importance'),
                                 value=c(35.2479,29.4679,22.3875,7.7451,4.2089,0.9427,23.2646,23.7045,34.5552,3.6263,10.6631,4.1863)
)

plt_feature_importance <- ggplot(data=feature_importance, aes(x=feature, y=value, fill=indicator))+
  geom_bar(stat='identity', position=position_dodge())+
  geom_text(aes(label=round(value)), vjust=0.5, hjust=1.1, color="black",
            position = position_dodge(0.9), size=2)+
  scale_fill_manual(limits=c('permutation importance','contribution'),
                    labels=c('Permutation importance','Contribution'),
                    values=c('#76b5c5','#eab676'))+
  labs(x='Feature', y="Percentage", fill='Measure')+
  scale_x_discrete(limits = rev(levels(feature_importance$feature)))+
  coord_flip()+
  theme_cowplot(font_size = 7, font_family = 'sans')+
  theme(text = element_text(size = 7, family='sans'))


plot(plt_feature_importance)

## Combine C with free space for B
bottom_row <- plot_grid(
  NA, plt_feature_importance,
  labels = c('b', 'c'), rel_widths = c(1, 1.5), label_size = 7
)
plot(bottom_row)

## Export panel leaving free space for B
tiff("./results/fig_model_performance/fig_model_performance.tiff", units="cm", width=18.3, height=10, res=600)
plot_grid(plt_dens, bottom_row, labels = c('a', ''), label_size = 7, ncol = 1, rel_heights = c(0.75,1))
dev.off()

# Scatterplot (B) to be added with a graphics program
library(plot3D)
presence_points <- maxent_model@presence
absence_points <- maxent_model@absence

tiff("./results/fig_model_performance/fig_3dscatter.tiff", units="cm", width=7.32, height=5.71, res=600, pointsize=7)
svg("./results/fig_model_performance/fig_3dscatter.svg", width=7.32, height=5.71, pointsize=7)
par(mar = c(0, 0, 0, 0))
plt_scatter <- scatter3D(x=absence_points$bio12, y=absence_points$bio15, z=absence_points$bio1,
                         col = 'black',
                         pch = 20,
                         cex = 0.15,
                         colkey = FALSE,
                         bty = "b2",
                         ticktype = "detailed",
                         #xlab = 'BIO12', ylab='BIO7',zlab = "BIO15",
                         xlab = '', ylab='',zlab = "",
                         cex.lab=1, cex.axis=1, line=2,
                         alpha=.2)

plt_scatter <- scatter3D(x=presence_points$bio12, y=presence_points$bio15, z=presence_points$bio1,
                         col = 'red',
                         pch = 20,0,
                         cex = 0.15,
                         add = TRUE)

dev.off()
